34 webapp improve

+++ run-your-c++-code-on-the-web/cli.cpp
@@ -4,10 +4,10 @@
 #include "newtonraphson.hpp"
 
 int main() {
-  double initial_guess = -4;
-  double tolerance = 0.001;
+  float initial_guess = -4;
+  float tolerance = 0.001;
   NewtonRaphson newtonraphson(tolerance);
-  double root = newtonraphson.solve(initial_guess);
+  float root = newtonraphson.solve(initial_guess);
 
   std::cout << "Function root is approximately at x = ";
   std::cout << std::fixed << std::setprecision(2) << root << std::endl;
